requirements.txt
torch
torchvision
matplotlib
numpy
comet-ml
python-dotenv
sklearn